desktop/requirements.txt
# Desktop model server dependencies — pinned versions, ZERO cloud SDKs
# All inference runs locally via llama-cpp-python or vllm
llama-cpp-python==0.3.9
uvicorn==0.34.2
fastapi==0.115.12
pydantic==2.11.1
